requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-google-genai>=2.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
streamlit>=1.38.0
google-generativeai>=0.8.0